pyproject.toml
[tool.poetry]
name = "ipif-hub"
version = "0.1.0"
description = ""
authors = ["Richard Hadden <richard.hadden@oeaw.ac.at>"]

[tool.poetry.dependencies]
python = "^3.10"
acdh-django-browsing = "1.0.0"
Django = ">=3.2,<4"
requests = ">=2.25"
django-haystack = "^3.1.1"
pysolr = "^3.9.0"
djangorestframework = "^3.13.1"
python-dateutil = "^2.8.2"
celery = "^5.2.3"
redis = "^4.2.0"
psycopg2-binary = "^2.9.3"
django-email-verification = "^0.2.2"
PyJWT = "^2.4.0"
jsonschema = "^4.8.0"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}
numpy = "^1.23.2"



[tool.poetry.dev-dependencies]
black = "^22.1.0"
ipython = "^8.1.1"
mypy = "^0.950"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.5"
djangorestframework-stubs = "^1.7.0"
types-python-dateutil = "^2.8.19"
celery-types = "^0.13.1"
ipykernel = "^6.15.1"
types-jsonschema = "^4.9.0"
flake8 = "^5.0.4"
flake8-pytest-style = "^1.6.0"
flake8-tidy-imports = "^4.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "djangobaseproject.settings"

[tool.isort]
profile = "black"